Move component to root as it's going to be used in other parts too
